refactor: 분리 SkipStaggerWrapper
